Correct two thread safety issues
a) jimpleTypesOfFieldOrMethodDescriptor caches the results in a
   thread-unsafe way, just disable the cache for now

b) Several methods of Scene which mutate the set of classes are accessed in
   multiple threads, synchronize them

Remove the cache in jimpleTypesOfFieldOrMethodDescriptor, which is used in
convertInvokeDynamicInsn (which can be run concurrently for multiple
classes/methods).

I'm not sure how important the cache is I'll have to investigate this more.
We could probably make it work if it is really necessary.